Rename lsst.butler -> lsst.daf.butler
