+++ SConstruct
@@ -1,4 +1,4 @@
 # -*- python -*-
 from lsst.sconsUtils import scripts
 
-scripts.BasicSConstruct("butler")
+scripts.BasicSConstruct("daf_butler")
