+++ python/lsst/__init__.py
@@ -1,0 +1,3 @@
+import pkgutil
+import lsstimport
+__path__ = pkgutil.extend_path(__path__, __name__)
