+++ python/lsst/daf/butler/butler.py
