+++ python/lsst/daf/butler/ci_hsc/__init__.py
