+++ python/lsst/daf/butler/core/config.py
