+++ python/lsst/daf/butler/core/datasets.py
