+++ python/lsst/daf/butler/core/datastore.py
