+++ python/lsst/daf/butler/core/fileDescriptor.py
