+++ python/lsst/daf/butler/core/formatter.py
