+++ python/lsst/daf/butler/core/location.py
