+++ python/lsst/daf/butler/core/quantum.py
