+++ python/lsst/daf/butler/core/regions.py
