+++ python/lsst/daf/butler/core/registry.py
