+++ python/lsst/daf/butler/core/schema.py
