+++ python/lsst/daf/butler/core/storageClass.py
