+++ python/lsst/daf/butler/core/units.py
