+++ python/lsst/daf/butler/core/utils.py
