+++ tests/test_hsc.py
@@ -23,14 +23,14 @@
 
 import unittest
 
-import lsst.butler.ci_hsc
+import lsst.daf.butler.ci_hsc
 
 
 class TestDataTestCase(unittest.TestCase):
     """A test case for the ci_hsc Registry test data."""
 
     def setUp(self):
-        self.db = lsst.butler.ci_hsc.ingest.run(":memory:", create=True)
+        self.db = lsst.daf.butler.ci_hsc.ingest.run(":memory:", create=True)
 
     def tearDown(self):
         self.db.close()
